# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LISAN MVP API
  version: 1.1.0
  description: Auth, Placement, Learning Path, Gamification, Subscriptions, and AI tutor MVP contracts.
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            displayName: { type: string }
            status: { type: string }
paths:
  /auth/register:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                displayName: { type: string }
                country: { type: string, minLength: 2, maxLength: 2 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair.
  /auth/logout:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204':
          description: Logged out
  /auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user
  /placement/tests/{languageCode}:
    get:
      tags: [Placement]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: languageCode
          required: true
          schema: { type: string, enum: [en, el] }
      responses:
        '200': { description: Placement test payload }
  /placement/submit:
    post:
      tags: [Placement]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [testId, answers]
              properties:
                testId: { type: string, format: uuid }
                answers:
                  type: array
                  items:
                    type: object
                    required: [questionId, optionId]
                    properties:
                      questionId: { type: string, format: uuid }
                      optionId: { type: string, format: uuid }
      responses:
        '201': { description: Placement result }
...
  /gamification/status:
    get:
      tags: [Gamification]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: XP and streak snapshot }
  /gamification/award:
    post:
      tags: [Gamification]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, points]
              properties:
                source: { type: string }
                points: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '201': { description: XP granted }
        '429': { description: Cooldown active or daily cap reached }
...
  /subscriptions/me:
    get:
      tags: [Subscriptions]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current subscription }
  /subscriptions/subscribe:
    post:
      tags: [Subscriptions]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planCode]
              properties:
                planCode: { type: string }
      responses:
        '201': { description: Subscription created }
...